+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* **\[docs\]** Add some minimal user documentation about `@Reference` and `@Inverse` annotations.
+
 ### [v0.109](https://github.com/arez/arez/tree/v0.109) (2018-10-08)
 [Full Changelog](https://github.com/arez/arez/compare/v0.108...v0.109)